Add files via upload
In this project of face detection
we have taken 3 entity : entity1 - Priyanka, person2: - Kangana, entity3 - stranger
In our model we have trained it by inputing 127 images of entity1 and 100 images of entity 2. And if either of the entity is not detected the model will label the face as stranger.
Here we have "haarcascase_haarcascade_frontalface_default" model to detect the faces and has use the training classifier used is "LBPHFaceRecognizer".
This model can detect the entities in single image to several images.